[issue-1215] Fix bug show popup instruction when staking nominate for Vara Network
