+++ src/screens/Home/Staking/InstructionModal/index.tsx
@@ -1,11 +1,63 @@
+import React, { RefObject } from 'react';
 import { SwModal, Button, Icon, Typography } from 'components/design-system-ui';
 import { deviceHeight } from 'constants/index';
 import { useSubWalletTheme } from 'hooks/useSubWalletTheme';
 import { Aperture, CheckCircle, Coins, Eye, PlusCircle, ThumbsUp } from 'phosphor-react-native';
 import { View, Text, Linking } from 'react-native';
 import { ScrollView } from 'react-native-gesture-handler';
+import { SWModalRefProps } from 'components/design-system-ui/modal/ModalBaseV2';
+import { InstructionItem } from 'screens/Home/Staking/InstructionModal/InstructionItem';
 
-export function InstructionModal({ setDetailModalVisible, modalRef, modalVisible, modalTitle, onPressStake }) {
+interface Props {
+  setDetailModalVisible: (value: boolean) => void;
+  modalVisible: boolean;
+  modalTitle: string;
+  onPressStake: () => void;
+  modalRef?: RefObject<SWModalRefProps>;
+}
+
+const INSTRUCTION_DATA = [
+  {
+    icon: Aperture,
+    color: 'rgba(230, 71, 142, 1)',
+    id: 'ins_1',
+    title: 'Mint an NFT before staking',
+    content:
+      "If you're participating in the Coinbase - Vara quests, be sure to mint an NFT before staking to qualify for Coinbase rewards.",
+  },
+  {
+    icon: Coins,
+    color: 'rgba(230, 220, 37, 1)',
+    id: 'state_2',
+    title: 'Unstake and withdraw',
+    content:
+      'Once staked, your funds will be locked. Unstake your funds anytime and withdraw after a 7-day period. Keep in mind that these actions are not automated and will incur network fees.',
+  },
+  {
+    icon: CheckCircle,
+    color: 'rgba(76, 234, 172, 1)',
+    id: 'state_3',
+    title: 'Keep your free balance',
+    content:
+      'Ensure that your free balance (transferrable balance) includes a minimum of 12 VARA to cover your existential deposit and network fees associated with staking, unstaking, and withdrawals.',
+  },
+  {
+    icon: Eye,
+    color: 'rgba(78, 180, 242, 1)',
+    id: 'state_4',
+    title: 'Track your stake',
+    content: 'Keep an eye on your stake periodically, as rewards and staking status can fluctuate over time.',
+  },
+  {
+    icon: ThumbsUp,
+    color: 'rgba(170, 218, 98, 1)',
+    id: 'state_5',
+    title: 'Select active pool',
+    content: 'It is recommended that you select an active pool. Check out the list of active pools in our FAQ.',
+  },
+];
+
+export function InstructionModal({ setDetailModalVisible, modalRef, modalVisible, modalTitle, onPressStake }: Props) {
   const theme = useSubWalletTheme().swThemes;
   const footer = () => {
     return (
@@ -36,47 +88,7 @@ export function InstructionModal({ setDetailModalVisible, modalRef, modalVisible
         footer={footer()}
         modalStyle={{ maxHeight: '90%' }}>
         <ScrollView showsVerticalScrollIndicator={false} style={{ maxHeight: deviceHeight * 0.55 }}>
-          {[
-            {
-              icon: Aperture,
-              color: 'rgba(230, 71, 142, 1)',
-              id: 'ins_1',
-              title: 'Mint an NFT before staking',
-              content:
-                "If you're participating in the Coinbase - Vara quests, be sure to mint an NFT before staking to qualify for Coinbase rewards.",
-            },
-            {
-              icon: Coins,
-              color: 'rgba(230, 220, 37, 1)',
-              id: 'state_2',
-              title: 'Unstake and withdraw',
-              content:
-                'Once staked, your funds will be locked. Unstake your funds anytime and withdraw after a 7-day period. Keep in mind that these actions are not automated and will incur network fees.',
-            },
-            {
-              icon: CheckCircle,
-              color: 'rgba(76, 234, 172, 1)',
-              id: 'state_3',
-              title: 'Keep your free balance',
-              content:
-                'Ensure that your free balance (transferrable balance) includes a minimum of 12 VARA to cover your existential deposit and network fees associated with staking, unstaking, and withdrawals.',
-            },
-            {
-              icon: Eye,
-              color: 'rgba(78, 180, 242, 1)',
-              id: 'state_4',
-              title: 'Track your stake',
-              content: 'Keep an eye on your stake periodically, as rewards and staking status can fluctuate over time.',
-            },
-            {
-              icon: ThumbsUp,
-              color: 'rgba(170, 218, 98, 1)',
-              id: 'state_5',
-              title: 'Select active pool',
-              content:
-                'It is recommended that you select an active pool. Check out the list of active pools in our FAQ.',
-            },
-          ].map(item => (
+          {INSTRUCTION_DATA.map(item => (
             <InstructionItem {...item} key={item.id} />
           ))}
         </ScrollView>
@@ -101,27 +113,5 @@ export function InstructionModal({ setDetailModalVisible, modalRef, modalVisible
         </Typography.Text>
       </SwModal>
     </>
-  );
-}
-
-function InstructionItem({ icon, color, title, content }) {
-  const theme = useSubWalletTheme().swThemes;
-  return (
-    <View
-      style={{
-        backgroundColor: 'rgba(26, 26, 26, 1)',
-        borderRadius: theme.borderRadius,
-        padding: 14,
-        marginBottom: theme.marginXS,
-        flexDirection: 'row',
-      }}>
-      <View style={{ justifyContent: 'center', paddingRight: 10 }}>
-        <Icon phosphorIcon={icon} iconColor={color} weight="fill" />
-      </View>
-      <View style={{ flex: 1 }}>
-        <Typography.Title style={{ color: 'white' }}>{title}</Typography.Title>
-        <Typography.Text style={{ color: theme.colorTextSecondary, width: '100%' }}>{content}</Typography.Text>
-      </View>
-    </View>
   );
 }
